+++ ase/cli/info.py
@@ -2,6 +2,7 @@ import platform
 import sys
 
 from ase.utils import import_module, FileNotFoundError
+from ase.utils import search_current_git_hash
 from ase.io.formats import filetype, all_formats, UnknownFileTypeError
 from ase.io.ulm import print_ulm_info
 from ase.io.bundletrajectory import print_bundletrajectory_info
@@ -52,8 +53,14 @@ def print_info():
         except ImportError:
             versions.append((name, 'no'))
         else:
-            versions.append((name + '-' + module.__version__,
+            # Search for git hash
+            githash = search_current_git_hash(module)
+            if githash is None:
+                githash = ''
+            else:
+                githash = '-{:.10}'.format(githash)
+            versions.append((name + '-' + module.__version__ + githash,
                             module.__file__.rsplit('/', 1)[0] + '/'))
 
     for a, b in versions:
-        print('{:16}{}'.format(a, b))
+        print('{:25}{}'.format(a, b))